Merges checkPostResponseJSONLDAccept and checkPostResponseAccepted into checkPostResponseCreated
checkPostResponseCreated passes on a 201 or 202 with valid JSON-LD.
Checks for 201 or 202 before checking to see if it SHOULD be rejected from the input form.
Extra if to look for location header only in case of 201.
checkPostResponseCreated contains fail messages specific to wrong status code or JSON-LD not accepted.
Also deletes some duplicate descriptions.